Rust changes to be more idiomatic while still keeping our identity
