+++ internal/policies/mount/adsys_mount/src/errors.rs
@@ -1,11 +1,7 @@
 use std::fmt::{Display, Formatter, Result};
 use thiserror::Error;
 
-/// Our own error type
-///
-/// It's not good practice to return errors that are not from your own API
-/// (unless they are from the std library, i.e. io or fmt errors).
-#[non_exhaustive] // Blocks match expressions without wildcards (_) for this enum.
+/// Represents error codes used by adsys_mount.
 #[derive(Debug, Error)]
 pub enum AdsysMountError {
     MountError = 1,
